remove imageViews from viewModels; update userDataProvider

+++ app/src/main/java/com/example/androidcourseshpp/data/dataProvider/UserDataProvider.kt
@@ -1,8 +1,8 @@
 package com.example.androidcourseshpp.data.dataProvider
 
 interface UserDataProvider {
-    fun saveUserServerId(userServerId: Long)
-    fun getUserServerId(): Long
+    fun saveUserServerId(userServerId: Int)
+    fun getUserServerId(): Int
     fun clearUserServerId()
     fun saveUserAvatarUrl(avatar: String)
     fun getUserAvatarUrl(): String
