+++ app/src/main/java/com/example/androidcourseshpp/data/dataProvider/UserDataProviderImpl.kt
@@ -7,7 +7,7 @@ import javax.inject.Inject
 const val USER_INFO_STORE = "userInfo"
 const val USER_AVATAR_URL = "userAvatar"
 const val USER_SERVER_ID = "userServerId"
-const val DEFAULT_ID_VALUE: Long = -1
+const val DEFAULT_ID_VALUE: Int = -1
 const val DEFAULT_AVATAR_VALUE = ""
 
 
@@ -16,15 +16,15 @@ class UserDataProviderImpl @Inject constructor(@DataProviderPref private val sha
 
     private val editor = sharedPref.edit()
 
-    override fun saveUserServerId(userServerId: Long) {
-        editor.putLong(USER_SERVER_ID, userServerId).apply()
+    override fun saveUserServerId(userServerId: Int) {
+        editor.putLong(USER_SERVER_ID, userServerId.toLong()).apply()
     }
 
     override fun getUserServerId() =
-        sharedPref.getLong(USER_SERVER_ID, DEFAULT_ID_VALUE)
+        sharedPref.getLong(USER_SERVER_ID, DEFAULT_ID_VALUE.toLong()).toInt()
 
     override fun clearUserServerId() {
-        editor.putLong(USER_SERVER_ID, DEFAULT_ID_VALUE).apply()
+        editor.putLong(USER_SERVER_ID, DEFAULT_ID_VALUE.toLong()).apply()
     }
 
     override fun saveUserAvatarUrl(avatar: String) {
