+++ app/src/main/java/com/example/androidcourseshpp/ui/BaseFragment.kt
@@ -74,7 +74,7 @@ abstract class BaseFragment<VBinding : ViewBinding>(
         }
     }
 
-    protected fun moveToUserProfileScreen(userServerId: Long) {
+    protected fun moveToUserProfileScreen(userServerId: Int) {
         val intent = Intent(requireContext(), MainActivity::class.java)
 
         intent.putExtra(USER_SERVER_ID, userServerId)
