+++ app/src/main/java/com/example/androidcourseshpp/ui/screens/auth/splash/SplashContract.kt
@@ -12,7 +12,7 @@ class SplashContract {
 
     sealed interface Effect : ViewEffect {
 
-        data class NavigateToUserProfileScreen(val userServerId: Long) : Effect
+        data class NavigateToUserProfileScreen(val userServerId: Int) : Effect
         data object NavigateToSignInScreen : Effect
     }
